pasketti-word/requirements-train.txt
torch
torchaudio
nemo_toolkit[asr]>=2.5.0
lightning
librosa
soundfile
jiwer
wandb
